updated all files package name and removed previous files
