+++ src/main/java/com/inventrik/digitalestore/DigitalEstoreApplication.java
@@ -1,4 +1,4 @@
-package com.example.digitalestore;
+package com.inventrik.digitalestore;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
