+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ server.port=8080
 
 # Logging Configuration
 logging.level.org.springframework=INFO
-logging.level.com.example.digitalestore=DEBUG
+logging.level.com.inventrik.digitalestore=DEBUG
 logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
